database context is now being new'd up every time it is used. a wrapper might need to be added in order to wrap proper tests around the repositories.
adding users table.
adding script to add users upon sucessful authentication with a third party.
returning a user in the register request.
only allowing unique user tokens to be registered.
adding AES encryption algorithm so that data in the database can be secured.
adding testing encryption key to the project, this needs to be properly secured for production.